teapot: Fix GCC warning

+++ bezier_teapot/teapot.cpp
@@ -594,7 +594,7 @@ void onDisplay()
     0                  // offset of first element
   );
   glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, ibo_teapot_cp_elements);
-  GLushort offset = 0;
+  size_t offset = 0;  // size_t instead of GLushort to fix 'warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]'
   for (int p = 0; p < TEAPOT_NB_PATCHES; p++)
     for (int i = 0; i < ORDER+1; i++, offset += (ORDER+1)*sizeof(GLushort))
       glDrawElements(GL_LINE_LOOP, ORDER+1, GL_UNSIGNED_SHORT, (GLvoid*)offset);
